Version 2.0006
The tab key now does word completion
outside of a math environment, or
produces a tab at the beginning of the
line or if the previous character is a
space.  Within a math environment, it takes you to the next set of
parentheses, or to the end of the
environment.  Other small fixes.